Revert "allow default export path to be overridden in config"
This reverts commit 6b251dc7c3d74961e8d24a52b0183ce4d572b2e9.
