+++ lib/config.js
@@ -14,6 +14,5 @@ export default {
 	reportingInterval: 5000,
 	reportingEndpoint: 'https://hub.bitclock.io',
 	reportingAPIVersion: 'v0',
-	__resolvedModule: null,
 	...JSON.parse(get(global, ['process', 'env', '__SECRET_BITCLOCK_CONFIG_JSON'], '{}'))
 };
