--- test/dummy-export.js
@@ -1,1 +1,0 @@
-export default { dummy: true };
